refactor: use `next-connect` in `/migrations` and `/status`
